SimpleInvoiceSystem/backend/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
pytz==2024.1
sqlparse==0.4.4
asgiref==3.7.1
